Proyecto/app/requirements.txt
flask
Wand
requests
python-dotenv